sketched in untested work push and pop [no ci]

+++ src/actions/work/actions.sh
@@ -343,15 +343,23 @@ work-report() {
   tput sgr0 # TODO: put this in the exit trap, too, I think.
 }
 
+# See 'liq help work resume'
 work-resume() {
+  eval "$(setSimpleOptions POP -- "$@")"
   local WORK_NAME
-  workUserSelectOne WORK_NAME '' true "$@"
+  if [[ -z "$POP" ]]; then
+    workUserSelectOne WORK_NAME '' true "$@"
 
-  if [[ -L "${LIQ_WORK_DB}/curr_work" ]]; then
-    if [[ "${LIQ_WORK_DB}/curr_work" -ef "${LIQ_WORK_DB}/${WORK_NAME}" ]]; then
-      echowarn "'$WORK_NAME' is already the current unit of work."
-      exit 0
+    if [[ -L "${LIQ_WORK_DB}/curr_work" ]]; then
+      if [[ "${LIQ_WORK_DB}/curr_work" -ef "${LIQ_WORK_DB}/${WORK_NAME}" ]]; then
+        echowarn "'$WORK_NAME' is already the current unit of work."
+        exit 0
+      fi
     fi
+  else
+    local PREV_WORK
+    PREV_WORK="$(ls ${LIQ_WORK_DB}/prev_work?? | sort --reverse | head -n 1)"
+    mv "$PREV_WORK" "${LIQ_WORK_DB}/curr_work"
   fi
 
   requireCleanRepos "${WORK_NAME}"
@@ -362,6 +370,7 @@ work-resume() {
   echo "Resumed '$WORK_NAME'."
 }
 
+# Alias for 'work-resume'
 work-join() { work-resume "$@"; }
 
 work-save() {
@@ -504,7 +513,7 @@ work-status() {
 }
 
 work-start() {
-  eval "$(setSimpleOptions ISSUES= -- "$@")"
+  eval "$(setSimpleOptions ISSUES= PUSH -- "$@")"
 
   local CURR_PROJECT ISSUES_URL
   if [[ -n "$BASE_DIR" ]]; then
@@ -535,7 +544,20 @@ work-start() {
   # https://github.com/Liquid-Labs/liq-cli/issues/14
 
   if [[ -L "${LIQ_WORK_DB}/curr_work" ]]; then
-    rm "${LIQ_WORK_DB}/curr_work"
+    if [[ -n "$PUSH" ]]; then
+      local PREV_WORK LAST NEXT
+      LAST='-1' # starts us at 0 after the undconditional +1
+      for PREV_WORK in $(ls ${LIQ_WORK_DB}/prev_work?? | sort); do
+        LAST=${i:$((${#i} - 2))}
+      done
+      NEXT=$(( $LAST + 1 ))
+      if (( $NEXT > 99 )); then
+        echoerrandexit "There are already 100 'pushed' units of work; limit reached."
+      fi
+      mv "${LIQ_WORK_DB}/curr_work" "${LIQ_WORK_DB}/prev_work$(printf '%02d' "${NEXT}")"
+    else
+      rm "${LIQ_WORK_DB}/curr_work"
+    fi
   fi
   touch "${LIQ_WORK_DB}/${WORK_BRANCH}"
   cd ${LIQ_WORK_DB} && ln -s "${WORK_BRANCH}" curr_work
@@ -631,6 +653,7 @@ work-submit() {
       requireCleanRepo "${IP}"
     fi
     # TODO: This is incorrect, we need to check IP; https://github.com/Liquid-Labs/liq-cli/issues/121
+    # TODO: might also be redundant with 'requireCleanRepo'...
     if ! work-status --pr-ready; then
       echoerrandexit "Local work branch not in sync with remote work branch. Try:\nliq work save --backup-only"
     fi
@@ -672,7 +695,7 @@ work-submit() {
       BASE_TARGET=$(git remote -v | grep '^upstream' | grep '(push)' | sed -E 's|.+[/:]([^/]+)/[^/]+$|\1|')
 
       local DESC
-      # recal, the first line is used in the 'summary' (title), the rest goes in the "description"
+      # recall, the first line is used in the 'summary' (title), the rest goes in the "description"
       DESC=$(cat <<EOF
 Merge ${WORK_BRANCH} to master
